+++ CHANGELOG.md
@@ -136,3 +136,4 @@
 - 2023-10-17: Pool qualifier, deals, session, gp, +1
 - 2023-10-18: Builder deals, session, qualifier, scheduler, +3
 - 2023-10-19: Wire deals, qualifier, llm
+- 2023-10-20: Features qualifier, deals